requirements.txt
flask
transformers
torch
scikit-learn
numpy
pandas
pymongo
python-dotenv
flask-cors
